Refactor to publish / subscribe pattern
Rename `Event` to `Topic` because it is actually
a topic from a publish / suscribe perspective.

Drop a lot of unit tests since the become obsolte due to
the new design. However, they are not replaced with
other unit tests, since there is alreadt a net of
integration tests in place.